can delete a facility now
it shows up as deleted from the facility list, but still can see if you do search facility. ATTEMPTED to do the add, but it doesnt work pls help. also i tried to do it so that the username is ADMIN, then they can click the manage button but its not rlly in the requirements so i just commented it out bc it doesnt work